Do not forward props uselessly
