+++ packages/gamut-icons/src/index.ts
@@ -1,3 +1,3 @@
 export * from './icons/regular';
 export * from './icons/mini';
-export * from './types';
+export { GamutIconProps } from './props';
